Added script to replace emoji names with emojis
I am very very tired of copy/pasting emojis over and over again, so here is a script to convert emoji names into the emojis themselves. I'd like to make things that take user input run this on that input so that I don't have to copy/paste so many emoji names.